refactor(firecracker): move shared substrate packages to projects/firecracker/substrate
