+++ bazel/images/BUILD
@@ -41,7 +41,6 @@ multirun(
     commands = [
         "//bazel/tools/hf2oci:image.push",
         "//bazel/tools/image:image.push",
-        "//projects/agent_platform/egress-proxy:image.push",
         "//projects/agent_platform/fc-agent-init:image.push",
         "//projects/agent_platform/fc-agentd/chart:chart.push",
         "//projects/agent_platform/fc-agentd/rootfs-builder:image.push",
@@ -50,6 +49,7 @@ multirun(
         "//projects/agent_platform/semgrep-guest:image.push",
         "//projects/agent_platform/semgrep-scand/chart:chart.push",
         "//projects/agent_platform/semgrep-scand:image.push",
+        "//projects/firecracker/substrate/egress-proxy:image.push",
         "//projects/grimoire/api:image.push",
         "//projects/grimoire/frontend:image.push",
         "//projects/grimoire/ws-gateway:image.push",
