+++ projects/agent_platform/fc-agent-init/cmd/BUILD
@@ -15,7 +15,7 @@ go_library(
         "//projects/agent_platform/fc-agent-init/internal/reconnect",
         "//projects/agent_platform/fc-agent-init/internal/vsockdial",
         "//projects/agent_platform/fc-agent-init/internal/watchdog",
-        "//projects/agent_platform/vsockproto",
+        "//projects/firecracker/substrate/vsockproto",
     ] + select({
         "@rules_go//go/platform:android": [
             "@org_golang_x_sys//unix",
