+++ projects/agent_platform/fc-agent-init/cmd/main.go
@@ -34,7 +34,7 @@ import (
 	"github.com/jomcgi/homelab/projects/agent_platform/fc-agent-init/internal/reconnect"
 	"github.com/jomcgi/homelab/projects/agent_platform/fc-agent-init/internal/vsockdial"
 	"github.com/jomcgi/homelab/projects/agent_platform/fc-agent-init/internal/watchdog"
-	"github.com/jomcgi/homelab/projects/agent_platform/vsockproto"
+	"github.com/jomcgi/homelab/projects/firecracker/substrate/vsockproto"
 )
 
 func main() {
