+++ projects/agent_platform/fc-agent-init/internal/idle/BUILD
@@ -5,12 +5,12 @@ go_library(
     srcs = ["idle.go"],
     importpath = "github.com/jomcgi/homelab/projects/agent_platform/fc-agent-init/internal/idle",
     visibility = ["//projects/agent_platform/fc-agent-init:__subpackages__"],
-    deps = ["//projects/agent_platform/vsockproto"],
+    deps = ["//projects/firecracker/substrate/vsockproto"],
 )
 
 go_test(
     name = "idle_test",
     srcs = ["idle_test.go"],
     embed = [":idle"],
-    deps = ["//projects/agent_platform/vsockproto"],
+    deps = ["//projects/firecracker/substrate/vsockproto"],
 )
