+++ projects/agent_platform/fc-agent-init/internal/idle/idle_test.go
@@ -4,7 +4,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/jomcgi/homelab/projects/agent_platform/vsockproto"
+	"github.com/jomcgi/homelab/projects/firecracker/substrate/vsockproto"
 )
 
 func TestNotIdleWhileCallInFlight(t *testing.T) {
