+++ projects/agent_platform/semgrep-guest-init/internal/lspdriver/lspdriver.go
@@ -24,7 +24,7 @@ import (
 	"sync"
 	"time"
 
-	"github.com/jomcgi/homelab/projects/agent_platform/vsockproto"
+	"github.com/jomcgi/homelab/projects/firecracker/substrate/vsockproto"
 )
 
 // Transport is the stdio byte seam to the LSP server. Recv carries server->client
