+++ projects/agent_platform/semgrep-guest-init/internal/scanserver/BUILD
@@ -10,7 +10,7 @@ go_library(
     importpath = "github.com/jomcgi/homelab/projects/agent_platform/semgrep-guest-init/internal/scanserver",
     visibility = ["//projects/agent_platform/semgrep-guest-init:__subpackages__"],
     deps = [
-        "//projects/agent_platform/vsockproto",
+        "//projects/firecracker/substrate/vsockproto",
     ] + select({
         "@rules_go//go/platform:android": [
             "@org_golang_x_sys//unix",
@@ -26,5 +26,5 @@ go_test(
     name = "scanserver_test",
     srcs = ["scanserver_test.go"],
     embed = [":scanserver"],
-    deps = ["//projects/agent_platform/vsockproto"],
+    deps = ["//projects/firecracker/substrate/vsockproto"],
 )
