+++ projects/agent_platform/semgrep-guest-init/internal/scanserver/scanserver_test.go
@@ -7,7 +7,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/jomcgi/homelab/projects/agent_platform/vsockproto"
+	"github.com/jomcgi/homelab/projects/firecracker/substrate/vsockproto"
 )
 
 // fakeScanner returns canned findings (or an error) for any request.
