+++ projects/firecracker/substrate/vsockproto/BUILD
@@ -3,7 +3,7 @@ load("@rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "vsockproto",
     srcs = ["proto.go"],
-    importpath = "github.com/jomcgi/homelab/projects/agent_platform/vsockproto",
+    importpath = "github.com/jomcgi/homelab/projects/firecracker/substrate/vsockproto",
     visibility = ["//visibility:public"],
 )
